Fixed an issue with a trailing slash on Windows
On Windows, given the following configuration, the update would fail:

{
  "type"     : "local",
  "local"    : "c:\\test\\test.txt",
  "remote"   : "d:\\example"
}

It would try to execute the following commandline:

xcopy /y /e /h c:\temp\test.txt d:\example/\

This is due to a trailing slash on the remoteFile variable which wasn't
getting removed. The fix is to remove any trailing slashes on the
remoteFile variable. This is safe to do, because we always append our own
trailing backslash.