Windows: Stop empty directories from being copied
Changed the xcopy switch from /e to /s. This prevents empty directories
from being copied. This is especially important if you are copying just a
single file. When you copy a single file with the /e flag, xcopy will
also create the entire directory structure even from directories that are
not related to the file you are copying. This takes extra time, and
probably isn't what you would expect to happen.